Move the Dockerfile to tests.
In practice, docker only allows a single Dockerfile per directory.  This
makes the directory that the Dockerfile is in analogous to a namespace
for your container.  It's important to place these containers in the
correct location.  In this case, I believe that this Dockerfile belongs
in the tests namespace and that we should keep the root project
namespace open for something that potentially "executes" the project in
a more general sense.